feat: 实现带图片验证码的登录功能
功能变更:
- 添加图片验证码生成接口，支持数字验证码
- 实现登录时验证码校验功能
- 增加验证码开关配置项，支持灵活启用/禁用

优化点:
- 完善 JWT Claims 结构体字段类型定义
- 新增 GetAuth 中间件函数，方便从上下文中获取用户信息
- 改进登录错误处理，统一返回"用户名或密码错误"防止用户名爆破

配置调整:
- 在 settings.yaml 中新增 site.login.captcha 配置项
- 创建 config_site.go 文件管理站点相关配置

+++ utils/jwts/enter.go
@@ -14,8 +14,8 @@ import (
 // go get github.com/golang-jwt/jwt/v5
 
 type Claims struct {
-	UserID uint `json:"userID"`
-	RoleID uint `json:"roleID"` //区分管理员和普通用户
+	UserID int  `json:"userID"`
+	RoleID int8 `json:"roleID"` //区分管理员和普通用户
 }
 type MyClaims struct { //千万不能存密码，因为jwt是编码不是加密
 	Claims
